Safari-611.1.10/LayoutTests/fast/canvas/2d.getTransform.newobject.html

<!DOCTYPE html>
<html>
<head>
<script src="../../resources/js-test-pre.js"></script>
</head>
<body>
<script>
description("The test to ensure that each call to getTransform returns a new object.");

var ctx = document.createElement("canvas").getContext("2d");
var transform1 = ctx.getTransform();
var transform2 = ctx.getTransform();

shouldNotBe("transform1", "transform2");
shouldBe("transform1.a", "transform2.a");
shouldBe("transform1.b", "transform2.b");
shouldBe("transform1.c", "transform2.c");
shouldBe("transform1.d", "transform2.d");
shouldBe("transform1.e", "transform2.e");
shouldBe("transform1.f", "transform2.f");

</script>
<script src="../../resources/js-test-post.js"></script>
</body>
</html>
